move table to common
