+++ src/common/enum.ts
@@ -1,0 +1,31 @@
+export enum s_p {
+  B = "\x1b[1m",
+  U = "\x1b[4m",
+  r = "\x1b[31m",
+  br = "\x1b[91m",
+  bg_r = "\x1b[41m",
+  g = "\x1b[32m",
+  bg = "\x1b[92m",
+  bg_g = "\x1b[42m",
+  y = "\x1b[33m",
+  by = "\x1b[93m",
+  bg_y = "\x1b[43m",
+}
+export enum s_s {
+  B = "\x1b[22m",
+  U = "\x1b[24m",
+  r = "\x1b[39m",
+  br = "\x1b[39m",
+  bg_r = "\x1b[49m",
+  g = "\x1b[39m",
+  bg = "\x1b[39m",
+  bg_g = "\x1b[49m",
+  y = "\x1b[39m",
+  by = "\x1b[39m",
+  bg_y = "\x1b[49m",
+}
+export enum t_e {
+  br = "🚨🚨🚨🚨",
+  by = "⚠️ ⚠️ ⚠️ ⚠️",
+  bg = "🧪🧪🧪🧪",
+}
