+++ src/common/mod.ts
@@ -1,1 +1,8 @@
 export { unique } from "trailmix/common/Array.ts";
+
+export { importDefault, validPath } from "trailmix/common/File.ts";
+
+// @deno-types="./File.d.ts"
+export type { FileExtension, ImportOptions } from "trailmix/common/File.d.ts";
+
+export { resetTable, testFunction } from "trailmix/common/table.ts";
