+++ src/deps.ts
@@ -1,4 +1,5 @@
 // watcher deps
+// export { bytesToUuid } from "uuid/_common.ts";
 export { v4 } from "uuid/mod.ts";
 export { extname, relative, resolve } from "path/mod.ts";
 export { exists, existsSync } from "fs/mod.ts";
@@ -18,3 +19,18 @@ export type { BaseHandler, FileHandler } from "log/handlers.ts";
 export type { LogConfig as stdLogConfig } from "log/mod.ts";
 export { getLevelByName, LogLevels } from "log/levels.ts";
 export { LogRecord } from "log/logger.ts";
+
+// test deps
+export {
+  assertEquals,
+  AssertionError,
+  assertMatch,
+  assertNotEquals,
+  assertObjectMatch,
+  assertStrictEquals,
+  assertThrows,
+  assertThrowsAsync,
+  unimplemented,
+} from "testing/asserts.ts";
+export { Cell, Row, Table } from "cliffy/table";
+export type { ICell, IRow, ITableOptions } from "cliffy/table";
