+++ test/common/mod.ts
@@ -1,2 +1,1 @@
 export * from "test/common/utils.ts";
-export * from "test/common/table.ts";
