--- test/deps.ts
@@ -1,13 +1,0 @@
-// test deps
-export {
-  assertEquals,
-  AssertionError,
-  assertMatch,
-  assertNotEquals,
-  assertObjectMatch,
-  assertStrictEquals,
-  assertThrows,
-  assertThrowsAsync,
-  unimplemented,
-} from "testing/asserts.ts";
-export { Cell, Row, Table } from "cliffy/table";
